src/requirements.txt
streamlit
openai
PyMuPDF
beautifulsoup4
requests
youtube_transcript_api
sentence-transformers
faiss-cpu
numpy